Move "shift" action (swap) inline in the slide complete event handler
