+++ src/actions/RequestReviewActions.tsx
@@ -8,7 +8,7 @@ import { ButtonsModal } from '../components/modals/ButtonsModal'
 import { Airship } from '../components/services/AirshipInstance'
 import { lstrings } from '../locales/strings'
 import { config } from '../theme/appConfig'
-import { RootState } from '../types/reduxTypes'
+import { ThunkAction } from '../types/reduxTypes'
 
 const SWAP_COUNT_DATA_FILE = 'swapCountData.json'
 const MANY_SWAPS_TO_TRIGGER_REQUEST = 3
@@ -42,34 +42,37 @@ const requestReview = async () => {
   }
 }
 
-export const updateSwapCount = async (state: RootState) => {
-  const { account } = state.core
-  let swapCountData
-  try {
-    const swapCountDataStr = await account.disklet.getText(SWAP_COUNT_DATA_FILE)
-    swapCountData = JSON.parse(swapCountDataStr)
-  } catch (e: any) {
-    // File needs init
-    swapCountData = {
-      swapCount: 0,
-      hasReviewBeenRequested: false
+export const updateSwapCount = (): ThunkAction<Promise<void>> => {
+  return async (_dispatch, getState) => {
+    const state = getState()
+    const { account } = state.core
+    let swapCountData
+    try {
+      const swapCountDataStr = await account.disklet.getText(SWAP_COUNT_DATA_FILE)
+      swapCountData = JSON.parse(swapCountDataStr)
+    } catch (e: any) {
+      // File needs init
+      swapCountData = {
+        swapCount: 0,
+        hasReviewBeenRequested: false
+      }
     }
-  }
 
-  const hasReviewBeenRequested = swapCountData.hasReviewBeenRequested
-  if (!hasReviewBeenRequested) {
-    let swapCount = parseInt(swapCountData.swapCount)
-    swapCount++
-    swapCountData.swapCount = swapCount
-    if (swapCount >= MANY_SWAPS_TO_TRIGGER_REQUEST) {
-      await requestReview()
-      swapCountData.hasReviewBeenRequested = true
+    const hasReviewBeenRequested = swapCountData.hasReviewBeenRequested
+    if (!hasReviewBeenRequested) {
+      let swapCount = parseInt(swapCountData.swapCount)
+      swapCount++
+      swapCountData.swapCount = swapCount
+      if (swapCount >= MANY_SWAPS_TO_TRIGGER_REQUEST) {
+        await requestReview()
+        swapCountData.hasReviewBeenRequested = true
+      }
+      const swapCountDataStr = JSON.stringify(swapCountData)
+      await account.disklet.setText(SWAP_COUNT_DATA_FILE, swapCountDataStr).catch(e => {
+        // Failure to write the swapCount file is tolerable since it just means the user won't be
+        //  asked to make a review
+        console.log(`RequestReviewActions.updateSwapCount: Error writing file ${SWAP_COUNT_DATA_FILE}:`, e)
+      })
     }
-    const swapCountDataStr = JSON.stringify(swapCountData)
-    await account.disklet.setText(SWAP_COUNT_DATA_FILE, swapCountDataStr).catch(e => {
-      // Failure to write the swapCount file is tolerable since it just means the user won't be
-      //  asked to make a review
-      console.log(`RequestReviewActions.updateSwapCount: Error writing file ${SWAP_COUNT_DATA_FILE}:`, e)
-    })
   }
 }
